+++ sdk/network/arm-network/CHANGELOG.md
@@ -1,15 +1,37 @@
 # Release History
+    
+## 31.0.0 (2023-03-29)
+    
+**Features**
 
-## 30.2.1 (Unreleased)
+  - Added Interface GroupByUserSession
+  - Added Interface GroupByVariable
+  - Added Interface PolicySettingsLogScrubbing
+  - Added Interface WebApplicationFirewallScrubbingRules
+  - Added Type Alias ApplicationGatewayFirewallRateLimitDuration
+  - Added Type Alias ApplicationGatewayFirewallUserSessionVariable
+  - Added Type Alias ScrubbingRuleEntryMatchOperator
+  - Added Type Alias ScrubbingRuleEntryMatchVariable
+  - Added Type Alias ScrubbingRuleEntryState
+  - Added Type Alias WebApplicationFirewallScrubbingState
+  - Interface PolicySettings has a new optional parameter logScrubbing
+  - Interface Subnet has a new optional parameter applicationGatewayIPConfigurations
+  - Interface WebApplicationFirewallCustomRule has a new optional parameter groupByUserSession
+  - Interface WebApplicationFirewallCustomRule has a new optional parameter rateLimitDuration
+  - Interface WebApplicationFirewallCustomRule has a new optional parameter rateLimitThreshold
+  - Added Enum KnownApplicationGatewayFirewallRateLimitDuration
+  - Added Enum KnownApplicationGatewayFirewallUserSessionVariable
+  - Added Enum KnownScrubbingRuleEntryMatchOperator
+  - Added Enum KnownScrubbingRuleEntryMatchVariable
+  - Added Enum KnownScrubbingRuleEntryState
+  - Added Enum KnownWebApplicationFirewallScrubbingState
+  - Enum KnownWebApplicationFirewallRuleType has a new value RateLimitRule
 
-### Features Added
+**Breaking Changes**
 
-### Breaking Changes
-
-### Bugs Fixed
-
-### Other Changes
-
+  - Interface Subnet no longer has parameter applicationGatewayIpConfigurations
+    
+    
 ## 30.2.0 (2023-03-07)
     
 **Features**